+++ live/core/src/plugins/virtual-output.ts
@@ -24,14 +24,11 @@ const createVirtualOutput = (
   context: OutputContext<Config>
 ): Output<Config> => {
   const group = new ld.Group({ noBorder: true });
-  const label = new ld.Label(`pixels:`);
-  group.addChild(label);
+  group.addChild(new ld.Label(`pixels:`));
 
-  const pixelsInput = new ld.TextInput('');
-  group.addChild(pixelsInput);
+  const pixelsInput = group.addChild(new ld.TextInput(''));
 
-  const update = new ld.Button('Update');
-  group.addChild(update);
+  const update = group.addChild(new ld.Button('Update'));
 
   let pixels: {
     px: Array<{
@@ -67,7 +64,7 @@ const createVirtualOutput = (
     }
     for (let i = 0; i < pixels.rects.length; i++) {
       const samplePixel = Math.round(
-        (i / (pixels.rects.length - 1)) * (pixels.px.length - 1)
+        (i / Math.max(1, pixels.rects.length - 1)) * (pixels.px.length - 1)
       );
       const px = pixels.px[samplePixel];
       const c = px.color.transition(RGBA_BLACK, x);